+++ test/gtest/scopes/internal/smartscopes/smartscopesproxy/smartscopesproxy_test.cpp
@@ -140,6 +140,11 @@ TEST_F(smartscopesproxytest, ss_registry)
         EXPECT_EQ(2, meta.settings_definitions()[1].get_dict()["displayValues"].get_array().size());
         EXPECT_EQ("Celsius", meta.settings_definitions()[1].get_dict()["displayValues"].get_array()[0].get_string());
         EXPECT_EQ("Fahrenheit", meta.settings_definitions()[1].get_dict()["displayValues"].get_array()[1].get_string());
+        ASSERT_EQ(4, meta.tags().size());
+        EXPECT_EQ("music", meta.tags()[0]);
+        EXPECT_EQ("video", meta.tags()[1]);
+        EXPECT_EQ("news", meta.tags()[2]);
+        EXPECT_EQ("games", meta.tags()[3]);
     }
 
     // non-existent scope (via mw)